Implement communication requests
Use Retrofit 2 API (http://square.github.io/retrofit/#api-declaration)
to implement the services that call the requests needed by the app
to communicate with the server via rest using the entities already
implemented in the app.

Also, implement the service generator to create a rest client
based on a to-be stablished url.

This implementation is based on the one that can be found at:
https://futurestud.io/tutorials/retrofit-getting-started-and-android-client